error out of script early of hadoop job fails.

+++ doit.sh
@@ -1,3 +1,4 @@
+clear
 cp /mnt/shared/TweetParse.java .
 hadoop com.sun.tools.javac.Main TweetParse.java
 result=$?
@@ -9,6 +10,11 @@ fi
 jar -cvf TweetParse.jar TweetParse*.class
 hdfs dfs -rm -f -r -skipTrash /user/root/output
 hadoop jar TweetParse.jar TweetParse /user/root/data/tweets.txt /user/root/output
+result=$?
+if [[ $result -ne 0 ]]
+then
+    return $result
+fi
 rm -f ./output.txt
 hdfs dfs -get /user/root/output/part-00000 ./output.txt
 cp ./output.txt /mnt/shared
